changed float to double for mm

+++ max-min/c/impl/mm.h
@@ -5,6 +5,6 @@
 #define MAX(X, Y) (((X) > (Y)) ? (X) : (Y))
 
 /* Computes the paths that maximize the minimum edge weight and stores their lengths in C. N is the number of nodes.*/
-int floydWarshall(float *C, int N);
+int floydWarshall(double *C, int N);
 
 #endif
