+++ max-min/c/impl/naive.c
@@ -3,10 +3,14 @@
 /*
 unoptimized version of the modified Floyd-Warshall maximin algorithm
 */
-int floydWarshall(float *C, int N) {
-    for (int k = 0; k < N; k++) {
-        for (int i = 0; i < N; i++) {
-            for (int j = 0; j < N; j++) {
+int floydWarshall(double *C, int N)
+{
+    for (int k = 0; k < N; k++)
+    {
+        for (int i = 0; i < N; i++)
+        {
+            for (int j = 0; j < N; j++)
+            {
                 C[i * N + j] = MAX(C[i * N + j], MIN(C[i * N + k], C[k * N + j]));
             }
         }
