+++ max-min/c/main.c
@@ -15,9 +15,12 @@
 #define FREQUENCY 2.3e9
 #define CALIBRATE
 
-void printMatrix(float *C, int N) {
-    for (int i = 0; i < N; i++) {
-        for (int j = 0; j < N; j++) {
+void printMatrix(double *C, int N)
+{
+    for (int i = 0; i < N; i++)
+    {
+        for (int j = 0; j < N; j++)
+        {
             fprintf(stderr, "%.2f, ", C[i * N + j]);
         }
         fprintf(stderr, "\n");
@@ -28,9 +31,11 @@ void printMatrix(float *C, int N) {
  * Runs the FW implementation once for testing purposes
  * Note that the matrix C is modified in-place
  */
-void ref_output(float *C, int N) {
+void ref_output(double *C, int N)
+{
     int err = floydWarshall(C, N);
-    if (err != 0) {
+    if (err != 0)
+    {
         printf("implementation reported an error\n");
         exit(1);
     }
@@ -40,10 +45,13 @@ void ref_output(float *C, int N) {
 Copies data from one matrix to another.
 Assumes matrices 'from' and 'to' point to separate memory locations.
 */
-void copyMatrix(float *from, float *to, int N) {
+void copyMatrix(double *from, double *to, int N)
+{
 #pragma ivdep
-    for (int i = 0; i < N; i++) {
-        for (int j = 0; j < N; j++) {
+    for (int i = 0; i < N; i++)
+    {
+        for (int j = 0; j < N; j++)
+        {
             to[i * N + j] = from[i * N + j];
         }
     }
@@ -59,7 +67,8 @@ void copyMatrix(float *from, float *to, int N) {
  *
  * The function returns the average number of cycles per run.
  */
-unsigned long long rdtsc(float *C, int N) {
+unsigned long long rdtsc(double *C, int N)
+{
     int i, num_runs;
     myInt64 cycles;
     myInt64 start;
@@ -72,9 +81,11 @@ unsigned long long rdtsc(float *C, int N) {
      * avoid measurements bias due to the timing overhead.
      */
 #ifdef CALIBRATE
-    while (num_runs < (1 << 14)) {
+    while (num_runs < (1 << 14))
+    {
         start = start_tsc();
-        for (i = 0; i < num_runs; ++i) {
+        for (i = 0; i < num_runs; ++i)
+        {
             floydWarshall(C, N);
         }
         cycles = stop_tsc(start);
@@ -97,7 +108,8 @@ unsigned long long rdtsc(float *C, int N) {
      */
 
     start = start_tsc();
-    for (i = 0; i < num_runs; ++i) {
+    for (i = 0; i < num_runs; ++i)
+    {
         floydWarshall(C, N);
     }
     cycles = stop_tsc(start) / num_runs;
@@ -106,24 +118,32 @@ unsigned long long rdtsc(float *C, int N) {
 }
 #endif
 
-void output_matrix(char *filename, float *C, int N) {
+void output_matrix(char *filename, double *C, int N)
+{
     fprintf(stderr, "outputting max-min matrix to %s...\n", filename);
     FILE *output_f = fopen(filename, "w+");
     fprintf(output_f, "%d\n", N);
-    for (int i = 0; i < N; i++) {
-        for (int j = 0; j < N; j++) {
+    for (int i = 0; i < N; i++)
+    {
+        for (int j = 0; j < N; j++)
+        {
             fprintf(output_f, "%.2f", C[i * N + j]);
-            if (j < N - 1) {
+            if (j < N - 1)
+            {
                 fputc(',', output_f);
-            } else {
+            }
+            else
+            {
                 fputc('\n', output_f);
             }
         }
     }
 }
 
-int main(int argc, char **argv) {
-    if (argc != 3) {
+int main(int argc, char **argv)
+{
+    if (argc != 3)
+    {
         fprintf(stderr, "incorrect number of arguments\n");
         fprintf(stderr, "call as: ./main input_filename output_filename\n");
         return -1;
@@ -131,31 +151,37 @@ int main(int argc, char **argv) {
 
     FILE *input_f = fopen(argv[1], "r");
 
-    int N;  // num nodes
-    if (fscanf(input_f, "%d ", &N) < 1) {
+    int N; // num nodes
+    if (fscanf(input_f, "%d ", &N) < 1)
+    {
         fprintf(stderr, "malformed input: couldn't match number N of vertices\n");
         return -1;
     }
 
     fprintf(stderr, "allocating memory...\n");
-    float *C = (float *)malloc(N * N * sizeof(float));
+    double *C = (double *)malloc(N * N * sizeof(double));
     fprintf(stderr, "parsing input matrix...\n");
-    for (int i = 0; i < N; i++) {
+    for (int i = 0; i < N; i++)
+    {
         char inputValue[100];
-        for (int j = 0; j < N; j++) {
+        for (int j = 0; j < N; j++)
+        {
             int numValues = fscanf(input_f, "%[^,\n]s", inputValue);
-            if (numValues == 1) {
+            if (numValues == 1)
+            {
                 C[i * N + j] = strtof(inputValue, NULL);
-            } else {
+            }
+            else
+            {
                 C[i * N + j] = 0.0;
             }
-            fgetc(input_f);  // skip ',' or '\n'
+            fgetc(input_f); // skip ',' or '\n'
         }
     }
     fclose(input_f);
 
-    float *D = (float *)malloc(N * N * sizeof(float));
-    memcpy(D, C, N * N * sizeof(float));
+    double *D = (double *)malloc(N * N * sizeof(double));
+    memcpy(D, C, N * N * sizeof(double));
     fprintf(stderr, "generating test output...\n");
     ref_output(D, N);
     char ref_output[256];
